fix: fixed onedark not loading properly
The reason why onedark won't load, throwing onedark#GetColors() as an
error, is that vimplug puts the autoload files in the wrong place.
Vimplug puts the items in autoload/ into autoload/onedark.vim/, which
can't be loaded because it's a directory.

A hack for this is to go into autoload/onedark.vim/autoload/, then move
all its contents to autoload/